Create release.sh
A script to merge your current branch into target branch and then switch back.

**Howto:**
Create and alias in mac or windows. For example 'deploy'
Example you are on branch: feature-a. Go to your terminal, cd to your project and type: `deploy development`. It will merge feature-a into development and then switch back to feature-a if there is no conflict.